Nashorn integration, i.e. extend sensors using javascript
This is probably rather slow and should NOT be used for actual production
environments. Though it might be useful for prototyping integrations or
for people who just want to get that sensor to do what they need to do
*now*.

An example integration is included, but a config option to disable nashorn
integration altogether is still missing.